Removed spaces in functions
There are spaces separating the function name from the parentheses, which I believe is an err (please correct me if I'm wrong). Also, is there a reason behind so many calls of file.ReadLine that return null? If not, let me know and I can go through and condense/optimize the code